Added a first validations to account management
