+++ apps/website/components/templates/settings/account-management/types.tsx
@@ -1,11 +1,11 @@
 import { z } from 'zod';
 
-export type AccountManagementProps = {
+export type emailProps = {
   emailAddress: string;
   code: string;
 };
 
-export const accountManagementSchema = z.object({
+export const emailSchema = z.object({
   email: z.string().email(),
-  code: z.string(),
+  code: z.string({ required_error: 'Code is required'}).min(1, { message: 'Code is required'}),
 });
