+++ apps/website/locales/en/settings.json
@@ -55,6 +55,10 @@
     "here": "here",
     "code-placeholder": "Code",
     "code-action": "Verify",
-    "code-send-again": "send code again"
+    "code-send-again": "send code again",
+    "email-sent": "Confirmation code sent to your email",
+    "email-verified": "Email successfully verified",
+    "email-error-sending": "There was an error sending the code",
+    "email-error-validating": "There was an error validating the code"
   }
 }
